refactor: 단순 패키지 구조 변경
