+++ src/main/java/com/st/eighteen_be/chat/service/impl/ChatMessageService.java
@@ -3,8 +3,8 @@ package com.st.eighteen_be.chat.service.impl;
 import com.st.eighteen_be.chat.model.collection.ChatMessageCollection;
 import com.st.eighteen_be.chat.model.dto.request.ChatMessageRequestDTO;
 import com.st.eighteen_be.chat.model.dto.response.ChatMessageResponseDTO;
-import com.st.eighteen_be.chat.repository.ChatMessageCollectionRepository;
-import com.st.eighteen_be.chat.repository.ChatroomInfoCollectionRepository;
+import com.st.eighteen_be.chat.repository.mongo.ChatMessageCollectionRepository;
+import com.st.eighteen_be.chat.repository.mongo.ChatroomInfoCollectionRepository;
 import com.st.eighteen_be.common.exception.ErrorCode;
 import com.st.eighteen_be.common.exception.sub_exceptions.data_exceptions.NotFoundException;
 import lombok.RequiredArgsConstructor;
