+++ src/test/java/com/st/eighteen_be/chat/service/facade/ChatroomFacadeTest.java
@@ -5,8 +5,8 @@ import com.st.eighteen_be.chat.model.collection.ChatroomInfoCollection;
 import com.st.eighteen_be.chat.model.dto.request.EnterChatRoomRequestDTO;
 import com.st.eighteen_be.chat.model.dto.response.ChatMessageResponseDTO;
 import com.st.eighteen_be.chat.model.vo.ChatroomType;
-import com.st.eighteen_be.chat.repository.ChatMessageCollectionRepository;
-import com.st.eighteen_be.chat.repository.ChatroomInfoCollectionRepository;
+import com.st.eighteen_be.chat.repository.mongo.ChatMessageCollectionRepository;
+import com.st.eighteen_be.chat.repository.mongo.ChatroomInfoCollectionRepository;
 import com.st.eighteen_be.chat.service.impl.ChatMessageService;
 import com.st.eighteen_be.chat.service.impl.ChatroomService;
 import com.st.eighteen_be.common.annotation.ServiceWithMongoDBTest;
